moved functions to services and removed config from test and development
