+++ services/githubService.js
@@ -1,5 +1,5 @@
 const axios = require('axios')
-const logger = require('../utils/logger')
+const logger = require('./logger')
 
 /**
  * Used for network calls
@@ -29,4 +29,22 @@ const fetch = async (url, method = 'get', params = null, data = null, headers = 
   }
 }
 
-module.exports = fetch
+/**
+ * Loops over an array of objects, takes a value corresponding to key provided and saves it in an array
+ *
+ * @param arrayOfObjects {Array} - Array of objects to loop over
+ * @param key {String} - Value corresponding to this key is saved
+ */
+
+const getNames = (arrayOfObjects, key) => {
+  const names = []
+  arrayOfObjects.forEach((object) => {
+    names.push(object[key])
+  })
+  return names
+}
+
+module.exports = {
+  fetch,
+  getNames
+}
